[visu] corrige bug de 2e appel de setSquareContent en cours de partie

+++ src/visu_src/models/MapData.cpp
@@ -6,7 +6,7 @@
 //   By: glourdel <glourdel@student.42.fr>          +#+  +:+       +#+        //
 //                                                +#+#+#+#+#+   +#+           //
 //   Created: 2014/06/03 15:25:43 by glourdel          #+#    #+#             //
-//   Updated: 2014/06/23 14:10:16 by glourdel         ###   ########.fr       //
+//   Updated: 2014/06/23 14:33:47 by glourdel         ###   ########.fr       //
 //                                                                            //
 // ************************************************************************** //
 
@@ -26,8 +26,8 @@ MapData::MapData(u32 width, u32 height, u32 textWidth, u32 textHeight)
 	  m_endOfGame(false)
 {
 	cout << "NEW MAPDATA : " << m_gridSize.Width << " x " << m_gridSize.Height << endl;
-	m_gridElemPSize.Width = m_texturePSize.Width / width;
-	m_gridElemPSize.Height = m_texturePSize.Height / height / 2;
+	m_gridElemPSize.Width = static_cast<u32>(static_cast<f32>(m_texturePSize.Width) / static_cast<f32>(width));
+	m_gridElemPSize.Height = static_cast<u32>(static_cast<f32>(m_texturePSize.Height) / static_cast<f32>(height) / 2.0f);
 // Init matrix
 	for (u32 i = 0; i < m_gridSize.Width; i++)
 	{
